update filtered points and deleted pointListController
